Split session into client/daemon sessions, start instructions
