+++ src/instructions.rs
@@ -1,0 +1,3 @@
+//! Farcaster instructions sent by client to daemon
+
+pub struct CreateNewSession {}
